refactor: 파일 이름, 구조 변경 (#46)
