+++ src/features/auth/index.ts
@@ -1,3 +1,2 @@
-export * from './auth.action';
 export * from './auth.api';
 export * from './auth.hook';
